楼盘条件查询问题（根据楼盘id查询->楼盘id集合取交集）

+++ src/main/java/com/jiubo/buildstore/bean/BuildingBean.java
@@ -236,7 +236,7 @@ public class BuildingBean implements Serializable {
     /**
      * 特价值
      */
-    private Integer specialOffer;
+    private BigDecimal specialOffer;
 
     /**
      * (1:近期开盘;2:不是近期开盘)
